feat(qk256): robust enhanced GGUF loader, GQA K/V slicer, diagnostics, and fail-fast flag
- loader (crates/bitnet-models/src/gguf_simple.rs)
  - Add orientation heuristics and diagnostics when detecting/handling QK256 tensors.
  - Compute/emit row stride, warn on suspicious aspect ratios, and surface errors instead of silently falling back.
  - Respect BITNET_DISABLE_MINIMAL_LOADER env var to fail-fast on enhanced-loader errors (used by CI/parity).
  - Emit final loader debug summary (hidden, heads, kv_heads, vocab, layers).
  - Improve config extraction tracing.

- quant (crates/bitnet-models/src/quant/i2s_qk256.rs)
  - Allow small alignment padding tolerance when validating packed QK256 data size (TOLERANCE=128).
  - Provide clearer error message on size/orientation mismatch.
  - Add regression test covering exact, padded, and out-of-tolerance sizes.

- weight mapping (crates/bitnet-models/src/weight_mapper.rs)
  - Return matched key from find_and_remove for better diagnostics.
  - Detect exporter quirk where K/V are emitted as [hidden, hidden] and slice to [kv_dim, hidden] for GQA.
  - Add extensive logging and an automated regression test (test_kv_slicing_for_gqa).
  - Emit derived/validation dimension logs and more helpful errors when shapes mismatch.

- transformer/models (crates/bitnet-models/src/transformer.rs & src/bitnet.rs)
  - Add informative tracing when constructing MHA layers and KVCache; validate divisibility of hidden/heads and heads/kv_heads.
  - Add debug prints in BitNetModel::from_gguf and parity eval to help parity/debug runs.

- parity/run & CI (scripts/parity_smoke.sh, .github/workflows/parity-proof.yml, docs/baselines/)
  - Set BITNET_DISABLE_MINIMAL_LOADER=1 in parity_smoke.sh and parity-proof workflow to enforce fail-fast behavior in CI.
  - Improve parity_smoke link check to safely find binary and run ldd.
  - Update parity baseline JSON to reflect rust-only/parity outcome.

- docs (docs/howto/use-qk256-models.md)
  - Document BITNET_DISABLE_MINIMAL_LOADER, loader behavior (enhanced vs minimal), and hidden×hidden K/V auto-fix.
  - Explain when to use the fail-fast flag (CI, parity, prod) with examples.

- misc/tests
  - Add unit/regression tests for QK256 tolerance and GQA K/V slicing to prevent silent fallback and ensure correct mapping.

Overall: make QK256 loading and weight mapping far more robust and observable, prevent silent fallback to the minimal loader by default in parity/CI, and provide diagnostics and tests to catch orientation/shape quirks early.

+++ tests/common/cross_validation/cpp_implementation.rs
@@ -964,7 +964,7 @@ mod tests {
             model_load_time_ms: 1000,
             tokenization_time_ms: 50,
             inference_time_ms: 200,
-            peak_memory: crate::BYTES_PER_MB, // 1MB
+            peak_memory: crate::BYTES_PER_MB as u32, // 1MB
             tokens_per_second: 100.5,
         };
